chore: rename src/ckb to src/fiber
