+++ src/fiber/network.rs
@@ -40,10 +40,10 @@ use super::key::blake2b_hash_with_salt;
 use super::types::{CFNMessage, Hash256, OpenChannel, Privkey, Pubkey};
 use super::CkbConfig;
 
-use crate::ckb::channel::{TxCollaborationCommand, TxUpdateCommand};
-use crate::ckb::types::TxSignatures;
 use crate::ckb_chain::contracts::{check_udt_script, is_udt_type_auto_accept};
 use crate::ckb_chain::{CkbChainMessage, FundingRequest, FundingTx, TraceTxRequest};
+use crate::fiber::channel::{TxCollaborationCommand, TxUpdateCommand};
+use crate::fiber::types::TxSignatures;
 use crate::{unwrap_or_return, Error};
 
 pub const CFN_PROTOCOL_ID: ProtocolId = ProtocolId::new(42);
