+++ src/lib.rs
@@ -5,8 +5,8 @@ pub mod ckb_chain;
 
 pub mod ldk;
 pub use ldk::{start_ldk, LdkConfig};
-pub mod ckb;
-pub use ckb::{start_ckb, CkbConfig, NetworkServiceEvent};
+pub mod fiber;
+pub use fiber::{start_ckb, CkbConfig, NetworkServiceEvent};
 pub mod cch;
 pub use cch::{start_cch, CchActor, CchConfig};
